// cypress/support/enums/fields-enum.ts
export enum EMPLOYEE_FIELDS_NAMES{
    EMPLOYEE_ID = 'Employee Id',
}


export enum COMMON_FIELDS_NAMES{
    USERNAME = 'Username',
    PASSWORD = 'Password',
    CONFIRM_PASSWORD = 'Confirm Password',
    FIRST_NAME = 'First Name',
    MIDDLE_NAME = 'Middle Name',
    LAST_NAME = 'Last Name',
    EMAIL = 'Email',
}

export enum LEAVE_FIELDS_NAMES{
    LEAVE_TYPE = 'Leave Type',
    ENTITLEMENTS = 'Entitlement',
    EMPLOYEE_NAME = 'Employee Name',
    FROM_DATE = 'From Date',
    LEAVE_STATUS = 'Show Leave with Status'
}

export enum JOB_FIELDS_NAMES{
    JOINED_DATE = 'Joined Date',
    JOB_TITLE = 'Job Title',
    JOB_SPECIFICATION = 'Job Specification',
    JOB_CATEGORY = 'Job Category',
    SUB_UNIT = 'Sub Unit',
    LOCATION = 'Location',
    EMPLOYMENT_STATUS = 'Employment Status'
}

export enum CANDIDATES_FIELDS_NAMES{
    CANDIDATE_NAME = 'Candidate Name',
    VACANCY = 'Vacancy',
    CONTACT_NUMBER = 'Contact Number',
    KEYWORDS = 'Keywords',
    DATE_OF_APPLICATION = 'Date of Application',
    NOTES = 'Notes'
}

export enum SCHEDULE_INTERVIEW_FIELDS_NAMES{
    INTERVIEW_TITLE = 'Interview Title',
    INTERVIEWER = 'Interviewer',
    DATE = 'Date',
    TIME = 'Time',
}